Move block graph definitions into their own file
